Update owl.video.js
I encountered a problem with a carousel containing youtube and vimeo videos. If a video is launched and you continue to slide, the video does not stop. Normaly the video is loaded in an iFrame which must be deleted after changing slides, but this no longer works (may be an incompatibility with recent versions of jQuery (here 3.4.1).

I found the problem: the iframe is supposed to be inserted in a div having the class "owl-video-frame", but the div in question is not added in the DOM.